requirements.txt
# ====================================================
# 🧠 Core ML + API Stack
# ====================================================
fastapi
uvicorn[standard]
python-multipart
pydantic

# ====================================================
# 🔥 Deep Learning
# ====================================================
torch
torchvision
torchaudio

# ====================================================
# 📊 Data Processing & Visualization
# ====================================================
numpy
pandas
scikit-learn
matplotlib
Pillow

# ====================================================
# 💬 NLP, Embeddings & Retrieval (Stage 2)
# ====================================================
sentence-transformers
transformers
langchain
faiss-cpu

# ====================================================
# ⚡ Optional Extras
# ====================================================
accelerate
streamlit